+++ Assignment2/src/main/java/ca/utoronto/utm/assignment2/paint/Square.java
@@ -85,7 +85,7 @@ public class Square implements Shape {
      * @return the stroke thickness of the Square
      */
     @Override
-    public int getThickness() {
+    public double getThickness() {
         return -1;
     }
