+++ Assignment2/src/main/java/ca/utoronto/utm/assignment2/paint/Squiggle.java
@@ -49,7 +49,7 @@ public class Squiggle implements Shape {
      * @return the stroke thickness of the Square
      */
     @Override
-    public int getThickness() {
+    public double getThickness() {
         return -1;
     }
